+++ src/Pagination/PaginationManipulator.php
@@ -4,7 +4,10 @@ namespace Nuwave\Lighthouse\Pagination;
 
 use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
+use GraphQL\Language\AST\TypeNode;
 use GraphQL\Language\Parser;
+use Illuminate\Container\Container;
+use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
@@ -71,9 +74,6 @@ class PaginationManipulator
         }
     }
 
-    /**
-     * Register connection with schema.
-     */
     protected function registerConnection(
         FieldDefinitionNode &$fieldDefinition,
         ObjectTypeDefinitionNode &$parentType,
@@ -94,13 +94,13 @@ class PaginationManipulator
         $connectionFieldName = addslashes(ConnectionField::class);
 
         $connectionType = Parser::objectTypeDefinition(/** @lang GraphQL */ <<<GRAPHQL
-            "A paginated list of $fieldTypeName edges."
-            type $connectionTypeName {
+            "A paginated list of {$fieldTypeName} edges."
+            type {$connectionTypeName} {
                 "Pagination information about the list of edges."
                 pageInfo: PageInfo! @field(resolver: "{$connectionFieldName}@pageInfoResolver")
 
                 "A list of $fieldTypeName edges."
-                edges: [$connectionEdgeName] @field(resolver: "{$connectionFieldName}@edgeResolver")
+                edges: [{$connectionEdgeName}!]! @field(resolver: "{$connectionFieldName}@edgeResolver")
             }
 GRAPHQL
         );
@@ -112,7 +112,7 @@ GRAPHQL
                 "An edge that contains a node of type $fieldTypeName and a cursor."
                 type $connectionEdgeName {
                     "The $fieldTypeName node."
-                    node: $fieldTypeName
+                    node: $fieldTypeName!
 
                     "A unique cursor that can be used for pagination."
                     cursor: String!
@@ -130,20 +130,15 @@ after: String
 GRAPHQL
         );
 
-        $fieldDefinition->type = Parser::namedType($connectionTypeName);
+        $fieldDefinition->type = $this->paginationResultType($connectionTypeName);
         $parentType->fields = ASTHelper::mergeUniqueNodeList($parentType->fields, [$fieldDefinition], true);
     }
 
-    /**
-     * Add the wrapping type for paginated results.
-     *
-     * This merges preexisting definitions to preserve maximum information.
-     */
     protected function addPaginationWrapperType(ObjectTypeDefinitionNode $objectType): void
     {
         $typeName = $objectType->name->value;
 
-        // If the type already exists, we use that instead
+        // Reuse existing types to preserve directives or other modifications made to it
         $existingType = $this->documentAST->types[$typeName] ?? null;
         if ($existingType !== null) {
             if (! $existingType instanceof ObjectTypeDefinitionNode) {
@@ -159,15 +154,12 @@ GRAPHQL
             $this->modelClass
             && ! ASTHelper::hasDirective($objectType, ModelDirective::NAME)
         ) {
-            $objectType->directives [] = Parser::constDirective('@model(class: "'.addslashes($this->modelClass).'")');
+            $objectType->directives [] = Parser::constDirective(/** @lang GraphQL */'@model(class: "'.addslashes($this->modelClass).'")');
         }
 
         $this->documentAST->setTypeDefinition($objectType);
     }
 
-    /**
-     * Register paginator with schema.
-     */
     protected function registerPaginator(
         FieldDefinitionNode &$fieldDefinition,
         ObjectTypeDefinitionNode &$parentType,
@@ -179,13 +171,13 @@ GRAPHQL
         $paginatorFieldClassName = addslashes(PaginatorField::class);
 
         $paginatorType = Parser::objectTypeDefinition(/** @lang GraphQL */ <<<GRAPHQL
-            "A paginated list of $fieldTypeName items."
-            type $paginatorTypeName {
+            "A paginated list of {$fieldTypeName} items."
+            type {$paginatorTypeName} {
                 "Pagination information about the list of items."
                 paginatorInfo: PaginatorInfo! @field(resolver: "{$paginatorFieldClassName}@paginatorInfoResolver")
 
-                "A list of $fieldTypeName items."
-                data: [$fieldTypeName!]! @field(resolver: "{$paginatorFieldClassName}@dataResolver")
+                "A list of {$fieldTypeName} items."
+                data: [{$fieldTypeName}!]! @field(resolver: "{$paginatorFieldClassName}@dataResolver")
             }
 GRAPHQL
         );
@@ -195,18 +187,15 @@ GRAPHQL
             self::countArgument($defaultCount, $maxCount)
         );
         $fieldDefinition->arguments [] = Parser::inputValueDefinition(/** @lang GraphQL */ <<<'GRAPHQL'
-"The offset from which elements are returned."
+"The offset from which items are returned."
 page: Int
 GRAPHQL
 );
 
-        $fieldDefinition->type = Parser::namedType($paginatorTypeName);
+        $fieldDefinition->type = $this->paginationResultType($paginatorTypeName);
         $parentType->fields = ASTHelper::mergeUniqueNodeList($parentType->fields, [$fieldDefinition], true);
     }
 
-    /**
-     * Register simple paginator with schema.
-     */
     protected function registerSimplePaginator(
         FieldDefinitionNode &$fieldDefinition,
         ObjectTypeDefinitionNode &$parentType,
@@ -218,13 +207,13 @@ GRAPHQL
         $paginatorFieldClassName = addslashes(SimplePaginatorField::class);
 
         $paginatorType = Parser::objectTypeDefinition(/** @lang GraphQL */ <<<GRAPHQL
-            "A paginated list of $fieldTypeName items."
-            type $paginatorTypeName {
+            "A paginated list of {$fieldTypeName} items."
+            type {$paginatorTypeName} {
                 "Pagination information about the list of items."
                 paginatorInfo: SimplePaginatorInfo! @field(resolver: "{$paginatorFieldClassName}@paginatorInfoResolver")
 
-                "A list of $fieldTypeName items."
-                data: [$fieldTypeName!]! @field(resolver: "{$paginatorFieldClassName}@dataResolver")
+                "A list of {$fieldTypeName} items."
+                data: [{$fieldTypeName}!]! @field(resolver: "{$paginatorFieldClassName}@dataResolver")
             }
 GRAPHQL
         );
@@ -234,12 +223,12 @@ GRAPHQL
             self::countArgument($defaultCount, $maxCount)
         );
         $fieldDefinition->arguments [] = Parser::inputValueDefinition(/** @lang GraphQL */ <<<'GRAPHQL'
-"The offset from which elements are returned."
+"The offset from which items are returned."
 page: Int
 GRAPHQL
         );
 
-        $fieldDefinition->type = Parser::namedType($paginatorTypeName);
+        $fieldDefinition->type = $this->paginationResultType($paginatorTypeName);
         $parentType->fields = ASTHelper::mergeUniqueNodeList($parentType->fields, [$fieldDefinition], true);
     }
 
@@ -248,7 +237,7 @@ GRAPHQL
      */
     protected static function countArgument(?int $defaultCount = null, ?int $maxCount = null): string
     {
-        $description = '"Limits number of fetched elements.';
+        $description = '"Limits number of fetched items.';
         if ($maxCount) {
             $description .= ' Maximum allowed value: '.$maxCount.'.';
         }
@@ -261,5 +250,26 @@ GRAPHQL
             );
 
         return $description.$definition;
+    }
+
+    /**
+     * @return \GraphQL\Language\AST\NamedTypeNode|\GraphQL\Language\AST\NonNullTypeNode
+     */
+    protected function paginationResultType(string $typeName): TypeNode
+    {
+        /** @var \Illuminate\Contracts\Config\Repository $config */
+        $config = Container::getInstance()->make(ConfigRepository::class);
+        $nonNull = $config->get('lighthouse.non_null_pagination_results')
+            ? '!'
+            : '';
+
+        /**
+         * We do not wrap the typename in [], so this will never be a ListOfTypeNode.
+         *
+         * @var \GraphQL\Language\AST\NamedTypeNode|\GraphQL\Language\AST\NonNullTypeNode $nonNullTypeNode
+         */
+        $nonNullTypeNode = Parser::typeReference(/** @lang GraphQL */ "{$typeName}{$nonNull}");
+
+        return $nonNullTypeNode;
     }
 }