+++ tests/Integration/Schema/Directives/HasManyDirectiveTest.php
@@ -2,7 +2,6 @@
 
 namespace Tests\Integration\Schema\Directives;
 
-use GraphQL\Error\Error;
 use GraphQL\Type\Definition\Type;
 use Illuminate\Support\Arr;
 use Nuwave\Lighthouse\Pagination\PaginationArgs;
@@ -288,22 +287,19 @@ class HasManyDirectiveTest extends DBTestCase
         }
         ';
 
-        $result = $this->graphQL(/** @lang GraphQL */ '
-        {
-            user {
-                tasks(first: 5) {
-                    data {
-                        id
+        $this
+            ->graphQL(/** @lang GraphQL */ '
+            {
+                user {
+                    tasks(first: 5) {
+                        data {
+                            id
+                        }
                     }
                 }
             }
-        }
-        ');
-
-        $this->assertSame(
-            PaginationArgs::requestedTooManyItems(3, 5),
-            $result->json('errors.0.message')
-        );
+            ')
+            ->assertGraphQLErrorMessage(PaginationArgs::requestedTooManyItems(3, 5));
     }
 
     public function testHandlesPaginationWithCountZero(): void
@@ -323,25 +319,20 @@ class HasManyDirectiveTest extends DBTestCase
         }
         ';
 
-        $this->graphQL(/** @lang GraphQL */ '
-        {
-            user {
-                id
-                tasks(first: 0) {
-                    data {
-                        id
+        $this
+            ->graphQL(/** @lang GraphQL */ '
+            {
+                user {
+                    id
+                    tasks(first: 0) {
+                        data {
+                            id
+                        }
                     }
                 }
             }
-        }
-        ')->assertJson([
-            'data' => [
-                'user' => [
-                    'id' => $this->user->id,
-                    'tasks' => null,
-                ],
-            ],
-        ])->assertGraphQLErrorCategory(Error::CATEGORY_GRAPHQL);
+            ')
+            ->assertGraphQLErrorMessage(PaginationArgs::requestedZeroOrLessItems(0));
     }
 
     public function testRelayTypeIsLimitedByMaxCountFromDirective(): void
@@ -488,8 +479,8 @@ class HasManyDirectiveTest extends DBTestCase
         );
 
         $user = $this->introspectType('User');
-
         $this->assertNotNull($user);
+
         /** @var array<string, mixed> $user */
         $tasks = Arr::first(
             $user['fields'],
@@ -499,7 +490,7 @@ class HasManyDirectiveTest extends DBTestCase
         );
         $this->assertSame(
             $expectedConnectionName,
-            $tasks['type']['name']
+            $tasks['type']/* TODO add back in in v6 ['ofType'] */['name']
         );
     }