+++ tests/Unit/Pagination/PaginateDirectiveTest.php
@@ -5,6 +5,7 @@ namespace Tests\Unit\Pagination;
 use GraphQL\Type\Definition\FieldArgument;
 use GraphQL\Type\Definition\FieldDefinition;
 use GraphQL\Type\Definition\ObjectType;
+use GraphQL\Utils\SchemaPrinter;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Pagination\PaginationArgs;
 use Nuwave\Lighthouse\Pagination\PaginationType;
@@ -12,6 +13,266 @@ use Tests\TestCase;
 
 class PaginateDirectiveTest extends TestCase
 {
+    public function testIncludesPaginationInfoObjectsInSchema(): void
+    {
+        $schema = $this->buildSchemaWithPlaceholderQuery();
+        $schemaString = SchemaPrinter::doPrint($schema);
+
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<'GRAPHQL'
+"""Information about pagination using a Relay style cursor connection."""
+type PageInfo {
+  """When paginating forwards, are there more items?"""
+  hasNextPage: Boolean!
+
+  """When paginating backwards, are there more items?"""
+  hasPreviousPage: Boolean!
+
+  """The cursor to continue paginating backwards."""
+  startCursor: String
+
+  """The cursor to continue paginating forwards."""
+  endCursor: String
+
+  """Total number of nodes in the paginated connection."""
+  total: Int!
+
+  """Number of nodes in the current page."""
+  count: Int!
+
+  """Index of the current page."""
+  currentPage: Int!
+
+  """Index of the last available page."""
+  lastPage: Int!
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<'GRAPHQL'
+"""Information about pagination using a fully featured paginator."""
+type PaginatorInfo {
+  """Number of items in the current page."""
+  count: Int!
+
+  """Index of the current page."""
+  currentPage: Int!
+
+  """Index of the first item in the current page."""
+  firstItem: Int
+
+  """Are there more pages after this one?"""
+  hasMorePages: Boolean!
+
+  """Index of the last item in the current page."""
+  lastItem: Int
+
+  """Index of the last available page."""
+  lastPage: Int!
+
+  """Number of items per page."""
+  perPage: Int!
+
+  """Number of total available items."""
+  total: Int!
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<'GRAPHQL'
+"""Information about pagination using a simple paginator."""
+type SimplePaginatorInfo {
+  """Number of items in the current page."""
+  count: Int!
+
+  """Index of the current page."""
+  currentPage: Int!
+
+  """Index of the first item in the current page."""
+  firstItem: Int
+
+  """Index of the last item in the current page."""
+  lastItem: Int
+
+  """Number of items per page."""
+  perPage: Int!
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+    }
+
+    /**
+     * @dataProvider nonNullPaginationResults
+     */
+    public function testManipulatesPaginator(bool $nonNullPaginationResults): void
+    {
+        config(['lighthouse.non_null_pagination_results' => $nonNullPaginationResults]);
+
+        $schema = $this->buildSchema(/** @lang GraphQL */ '
+        type User {
+            id: ID!
+        }
+
+        type Query {
+            users: [User!]! @paginate
+        }
+        ');
+        $schemaString = SchemaPrinter::doPrint($schema);
+
+        $nonNull = $nonNullPaginationResults
+            ? '!'
+            : '';
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<GRAPHQL
+type Query {
+  users(
+    """Limits number of fetched items."""
+    first: Int!
+
+    """The offset from which items are returned."""
+    page: Int
+  ): UserPaginator{$nonNull}
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<'GRAPHQL'
+"""A paginated list of User items."""
+type UserPaginator {
+  """Pagination information about the list of items."""
+  paginatorInfo: PaginatorInfo!
+
+  """A list of User items."""
+  data: [User!]!
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+    }
+
+    /**
+     * @dataProvider nonNullPaginationResults
+     */
+    public function testManipulatesSimplePaginator(bool $nonNullPaginationResults): void
+    {
+        config(['lighthouse.non_null_pagination_results' => $nonNullPaginationResults]);
+
+        $schema = $this->buildSchema(/** @lang GraphQL */ '
+        type User {
+            id: ID!
+        }
+
+        type Query {
+            users: [User!]! @paginate(type: SIMPLE)
+        }
+        ');
+        $schemaString = SchemaPrinter::doPrint($schema);
+
+        $nonNull = $nonNullPaginationResults
+            ? '!'
+            : '';
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<GRAPHQL
+type Query {
+  users(
+    """Limits number of fetched items."""
+    first: Int!
+
+    """The offset from which items are returned."""
+    page: Int
+  ): UserSimplePaginator{$nonNull}
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<'GRAPHQL'
+"""A paginated list of User items."""
+type UserSimplePaginator {
+  """Pagination information about the list of items."""
+  paginatorInfo: SimplePaginatorInfo!
+
+  """A list of User items."""
+  data: [User!]!
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+    }
+
+    /**
+     * @dataProvider nonNullPaginationResults
+     */
+    public function testManipulatesConnection(bool $nonNullPaginationResults): void
+    {
+        config(['lighthouse.non_null_pagination_results' => $nonNullPaginationResults]);
+
+        $schema = $this->buildSchema(/** @lang GraphQL */ '
+        type User {
+            id: ID!
+        }
+
+        type Query {
+            users: [User!]! @paginate(type: CONNECTION)
+        }
+        ');
+        $schemaString = SchemaPrinter::doPrint($schema);
+
+        $nonNull = $nonNullPaginationResults
+            ? '!'
+            : '';
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<GRAPHQL
+type Query {
+  users(
+    """Limits number of fetched items."""
+    first: Int!
+
+    """A cursor after which elements are returned."""
+    after: String
+  ): UserConnection{$nonNull}
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<'GRAPHQL'
+"""A paginated list of User edges."""
+type UserConnection {
+  """Pagination information about the list of edges."""
+  pageInfo: PageInfo!
+
+  """A list of User edges."""
+  edges: [UserEdge!]!
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+
+        $this->assertStringContainsString(/** @lang GraphQL */ <<<'GRAPHQL'
+"""An edge that contains a node of type User and a cursor."""
+type UserEdge {
+  """The User node."""
+  node: User!
+
+  """A unique cursor that can be used for pagination."""
+  cursor: String!
+}
+GRAPHQL
+            ,
+            $schemaString
+        );
+    }
+
     public function testAliasRelayToConnection(): void
     {
         $connection = $this->getConnectionQueryField(PaginationType::CONNECTION);
@@ -129,7 +390,7 @@ class PaginateDirectiveTest extends TestCase
         $this->assertInstanceOf(FieldArgument::class, $defaultPaginatedAmountArg);
         /** @var \GraphQL\Type\Definition\FieldArgument $defaultPaginatedAmountArg */
         $this->assertSame(
-            'Limits number of fetched elements. Maximum allowed value: 5.',
+            'Limits number of fetched items. Maximum allowed value: 5.',
             $defaultPaginatedAmountArg->description
         );
 
@@ -140,7 +401,7 @@ class PaginateDirectiveTest extends TestCase
         $this->assertInstanceOf(FieldArgument::class, $defaultRelayFirstArg);
         /** @var \GraphQL\Type\Definition\FieldArgument $defaultRelayFirstArg */
         $this->assertSame(
-            'Limits number of fetched elements. Maximum allowed value: 5.',
+            'Limits number of fetched items. Maximum allowed value: 5.',
             $defaultRelayFirstArg->description
         );
 
@@ -151,7 +412,7 @@ class PaginateDirectiveTest extends TestCase
         $this->assertInstanceOf(FieldArgument::class, $defaultSimpleFirstArg);
         /** @var \GraphQL\Type\Definition\FieldArgument $defaultSimpleFirstArg */
         $this->assertSame(
-            'Limits number of fetched elements. Maximum allowed value: 5.',
+            'Limits number of fetched items. Maximum allowed value: 5.',
             $defaultSimpleFirstArg->description
         );
 
@@ -162,7 +423,7 @@ class PaginateDirectiveTest extends TestCase
         $this->assertInstanceOf(FieldArgument::class, $customPaginatedAmountArg);
         /** @var \GraphQL\Type\Definition\FieldArgument $customPaginatedAmountArg */
         $this->assertSame(
-            'Limits number of fetched elements. Maximum allowed value: 10.',
+            'Limits number of fetched items. Maximum allowed value: 10.',
             $customPaginatedAmountArg->description
         );
 
@@ -173,7 +434,7 @@ class PaginateDirectiveTest extends TestCase
         $this->assertInstanceOf(FieldArgument::class, $customRelayFirstArg);
         /** @var \GraphQL\Type\Definition\FieldArgument $customRelayFirstArg */
         $this->assertSame(
-            'Limits number of fetched elements. Maximum allowed value: 10.',
+            'Limits number of fetched items. Maximum allowed value: 10.',
             $customRelayFirstArg->description
         );
 
@@ -184,7 +445,7 @@ class PaginateDirectiveTest extends TestCase
         $this->assertInstanceOf(FieldArgument::class, $customSimpleFirstArg);
         /** @var \GraphQL\Type\Definition\FieldArgument $customSimpleFirstArg */
         $this->assertSame(
-            'Limits number of fetched elements. Maximum allowed value: 10.',
+            'Limits number of fetched items. Maximum allowed value: 10.',
             $customSimpleFirstArg->description
         );
     }
@@ -290,7 +551,10 @@ class PaginateDirectiveTest extends TestCase
         );
     }
 
-    public function testThrowsWhenPaginationWithCountZeroIsRequested(): void
+    /**
+     * @dataProvider nonNullPaginationResults
+     */
+    public function testThrowsWhenPaginationWithCountZeroIsRequested(bool $nonNullPaginationResults): void
     {
         $this->schema = /** @lang GraphQL */ '
         type User {
@@ -303,25 +567,36 @@ class PaginateDirectiveTest extends TestCase
         }
         ';
 
-        $result = $this->graphQL(/** @lang GraphQL */ '
-        {
-            users(first: 0) {
-                data {
-                    id
+        $result = $this
+            ->graphQL(/** @lang GraphQL */ '
+            {
+                users(first: 0) {
+                    data {
+                        id
+                    }
                 }
             }
-        }
-        ')
-        ->assertJson([
-            'data' => [
-                'users' => null,
-            ],
-        ]);
+            ')
+            ->assertGraphQLErrorMessage(PaginationArgs::requestedZeroOrLessItems(0));
 
-        $this->assertSame(
-            PaginationArgs::requestedZeroOrLessItems(0),
-            $result->json('errors.0.message')
-        );
+        if (! $nonNullPaginationResults) {
+            $result->assertJson([
+                'data' => [
+                    'users' => null,
+                ],
+            ]);
+        }
+    }
+
+    /**
+     * @return array<int, array{bool}>
+     */
+    public function nonNullPaginationResults(): array
+    {
+        return [
+            [true],
+            [false],
+        ];
     }
 
     public function testDoesNotRequireModelWhenUsingBuilder(): void